Fixes minetest.svg
In the scalable Minetest icon I designed, the black part in the middle of the screen meant to represent the black cave of the original icon, but you made it green. I think it would be best to revert it back to black, to follow the original icon's intention.

Also fixed the gradient artifacts from your edit.