install.sh: Fix read-only filesystem detection & use proper local system location for icon install
`-w` bash built-in & greping for `w` in `ls -ld` don't work properly, so try to make a temporary file. If it's made successfully, delete it & return code 1 for `check_read_only` function, otherwise return 0.

`/usr/local/` is a symlink to `/var/usrlocal/` in Fedora Atomic only, so use `/usr/local/` to be more compatible.